% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{suf_stat}
\alias{suf_stat}
\title{Calculates sufficient statistics for the variance decomposition}
\usage{
suf_stat(y, x, wgt = rep(1, length(y)), moment = "skewness")
}
\description{
Calculates sufficient statistics for the variance decomposition
}
