% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{suf_stat.var}
\alias{suf_stat.var}
\title{Calculates sufficient statistics for the variance decomposition}
\usage{
suf_stat.var(y, x, wgt = rep(1, length(y)))
}
\description{
Calculates sufficient statistics for the variance decomposition
}
