% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.vs_decomp}
\alias{summary.vs_decomp}
\title{Summary Method for 'vs_decomp' Objects}
\usage{
\method{summary}{vs_decomp}(object, r = 3, sum.comp = NULL, ...)
}
\arguments{
\item{object}{a 'vs_decomp' object.}

\item{r}{a number of decimal places to use.}

\item{sum.comp}{a vector of up to 3 components to summarize. can be either character
with component names or numeric with component indices. other
components not specified by sum.comp are summed to one additional
component. default is to summarize the first 3 components.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
No return value.
}
\description{
Summary Method for 'vs_decomp' Objects
}
