% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness_functions.R
\name{wtd_skew}
\alias{wtd_skew}
\title{Weighted Skewness}
\usage{
wtd_skew(x, wgt = rep(1, length(x)))
}
\arguments{
\item{x}{a numeric vector}

\item{wgt}{an optional vector of weights.}
}
\value{
scalar
}
\description{
computes weighted version of the skewness estimator.
}
