% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_projection.R
\name{linear_projection}
\alias{linear_projection}
\title{Linear Projection}
\usage{
linear_projection(y, X.list, data, wgt = rep(1, length(y)), comp.names = NULL)
}
\arguments{
\item{y}{a character specifing the name of the outcome variable (e.g. "wage").}

\item{X.list}{a list containing the components}

\item{data}{a data frame with all the variables specified in X.list and y.}

\item{wgt}{an optional vector of weights.}

\item{comp.names}{an optional vector specifing name for each component. should be the
same length as X.list.}
}
\description{
this function projects the standardized version of y onto all the variables in X.list
(so that \eqn{Y = \beta*X + \epsilon})
and returns
}
