% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vs_decomp}
\alias{plot.vs_decomp}
\title{Calculates sufficient statistics for the variance decomposition}
\usage{
\method{plot}{vs_decomp}(object, ylim = NULL)
}
\description{
Calculates sufficient statistics for the variance decomposition
}
