% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vs_decomp}
\alias{plot.vs_decomp}
\title{Plot Method for 'vs_decomp' Objects}
\usage{
\method{plot}{vs_decomp}(x, plot.comp = NULL, fill.colors = NULL, abs.terms = T, ...)
}
\arguments{
\item{x}{a 'vs_decomp' object.}

\item{plot.comp}{a vector of up to 3 components to plot. can be either character
with component names or numeric with component indices. other
components not specified by plot.comp are summed to one additional
component. default is to plot the first 3 components.}

\item{fill.colors}{colors to fill the areas. see \code{\link[ggplot2]{scale_fill_manual}}
for more details.}

\item{abs.terms}{whether to plot absolute terms. default is TRUE.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot Method for 'vs_decomp' Objects
}
