% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_decomp.R
\name{vs_decomp}
\alias{vs_decomp}
\title{Skewness and Variance Decomposition}
\usage{
vs_decomp(
  y = NULL,
  X,
  wgt = rep(1, nrow(X)),
  moment = "skewness",
  year = rep(1, nrow(X))
)
}
\arguments{
\item{y}{an outcome vector (necessary only when decomposing by one variable).}

\item{X}{when decomposing by one variable, a matrix or data frame containing the variable
for the decomposition. when decomposing by more than one variable, a matrix or
data frame containing the linear components of y, calculated by
\code{\link{linear_projection}}.}

\item{wgt}{an optional vector of weights.}

\item{moment}{the moment on which the decomposition method is applied.
either "variance" (second moment) or "skewness" (third moment).}

\item{year}{an optional vector of years. if provided, the decomposition is calculated by year.
otherwise, the decomposition is calculated for the whole sample.}
}
\value{

}
\description{
decompose the skewness or the variance of an outcome vector into independent components,
either using one or many variables.
}
