% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_decomp.R
\name{vs_decomp}
\alias{vs_decomp}
\title{Variance and Skewness Decomposition}
\usage{
vs_decomp(
  y,
  X,
  wgt = rep(1, length(y)),
  moment = "skewness",
  year = rep(1, length(y))
)
}
\arguments{
\item{y}{an outcome vector (must be numeric without NA's).}

\item{X}{}

\item{wgt}{an optional vector of weights.}

\item{moment}{the moment on which the decomposition method is applied.
either "variance" (second moment) or "skewness" (third moment).}

\item{year}{}
}
\description{
Variance and Skewness Decomposition
}
