% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_decomp.R
\name{vs_decomp}
\alias{vs_decomp}
\title{Variance and Skewness Decomposition}
\usage{
vs_decomp(
  y,
  X,
  wgt = rep(1, length(y)),
  moment = "skewness",
  year = rep(1, length(y))
)
}
\arguments{
\item{y}{an outcome vector (must be numeric without NA's).}

\item{X}{matrix or data frame containing variables for the decomposition. Usually, if decomposition by
more than one variable is desired, X should be the output of \code{\link{linear_projection}}.}

\item{wgt}{an optional vector of weights.}

\item{moment}{the moment on which the decomposition method is applied.
either "variance" (second moment) or "skewness" (third moment).}

\item{year}{an optional vector of years. if provided, the decomposition is calculated by year.
otherwise, the decomposition is calculated for the whole sample.}
}
\description{
Variance and Skewness Decomposition
}
